Add gas logic for move call and module publish

+++ fastx_programmability/adapter/src/adapter.rs
@@ -10,6 +10,10 @@ use fastx_types::{
         TX_CONTEXT_MODULE_NAME, TX_CONTEXT_STRUCT_NAME,
     },
     error::{FastPayError, FastPayResult},
+    gas::{
+        calculate_module_publish_cost, calculate_object_creation_cost,
+        calculate_object_deletion_refund, deduct_gas,
+    },
     object::{Data, MoveObject, Object},
     storage::Storage,
 };
@@ -46,7 +50,8 @@ pub fn execute<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error =
     type_args: Vec<TypeTag>,
     object_args: Vec<Object>,
     pure_args: Vec<Vec<u8>>,
-    gas_budget: Option<u64>,
+    gas_budget: u64,
+    mut gas_object: Object,
     ctx: TxContext,
 ) -> Result<(), FastPayError> {
     let TypeCheckSuccess {
@@ -65,8 +70,9 @@ pub fn execute<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error =
 
     let vm = MoveVM::new(natives)
         .expect("VM creation only fails if natives are invalid, and we created the natives");
+    // TODO: Update Move gas constants to reflect the gas fee on fastx.
     let mut gas_status =
-        get_gas_status(gas_budget).map_err(|e| FastPayError::GasBudgetTooHigh {
+        get_gas_status(Some(gas_budget)).map_err(|e| FastPayError::GasBudgetTooHigh {
             error: e.to_string(),
         })?;
     let session = vm.new_session(state_view);
@@ -82,7 +88,7 @@ pub fn execute<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error =
             events,
             return_values,
             mutable_ref_values,
-            gas_used: _, // TODO(https://github.com/MystenLabs/fastnft/issues/45): charge
+            gas_used,
         } => {
             // we already checked that the function had no return types in resolve_and_type_check--it should
             // also not return any values at runtime
@@ -94,12 +100,27 @@ pub fn execute<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error =
             let mutable_refs = mutable_ref_objects
                 .into_iter()
                 .zip(mutable_ref_values.into_iter());
-            process_successful_execution(state_view, by_value_objects, mutable_refs, events)?;
+            process_successful_execution(
+                state_view,
+                by_value_objects,
+                mutable_refs,
+                events,
+                gas_object,
+                gas_used,
+                gas_budget,
+            )?;
             Ok(())
         }
-        ExecutionResult::Fail { error, gas_used: _ } => Err(FastPayError::AbortedExecution {
-            error: error.to_string(),
-        }),
+        ExecutionResult::Fail { error, gas_used } => {
+            // Need to deduct gas even if the execution failed.
+            deduct_gas(&mut gas_object, gas_used as i128)?;
+            state_view.write_object(gas_object);
+            // TODO: Keep track the gas deducted so that we could give them to participants.
+
+            Err(FastPayError::AbortedExecution {
+                error: error.to_string(),
+            })
+        }
     }
 }
 
@@ -108,12 +129,17 @@ pub fn publish<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error =
     module_bytes: Vec<Vec<u8>>,
     sender: FastPayAddress,
     ctx: &mut TxContext,
+    mut gas_object: Object,
 ) -> Result<Vec<ObjectRef>, FastPayError> {
     if module_bytes.is_empty() {
         return Err(FastPayError::ModulePublishFailure {
             error: "Publishing empty list of modules".to_string(),
         });
     }
+    let gas_cost = calculate_module_publish_cost(&module_bytes);
+    deduct_gas(&mut gas_object, gas_cost as i128)?;
+    state_view.write_object(gas_object);
+    // TODO: Keep track the gas deducted so that we could give them to participants.
 
     let mut modules = module_bytes
         .iter()
@@ -200,6 +226,9 @@ fn process_successful_execution<
     mut by_value_objects: BTreeMap<ObjectID, Object>,
     mutable_refs: impl Iterator<Item = (Object, Vec<u8>)>,
     events: Vec<Event>,
+    mut gas_object: Object,
+    mut gas_used: u64,
+    gas_budget: u64,
 ) -> Result<(), FastPayError> {
     for (mut obj, new_contents) in mutable_refs {
         match &mut obj.data {
@@ -241,7 +270,9 @@ fn process_successful_execution<
                         old_object.next_sequence_number = sequence_number;
                         old_object
                     } else {
-                        Object::new_move(move_obj, recipient, SequenceNumber::new())
+                        let obj = Object::new_move(move_obj, recipient, SequenceNumber::new());
+                        gas_used += calculate_object_creation_cost(&obj);
+                        obj
                     };
                     new_object.owner = recipient;
                     state_view.write_object(new_object);
@@ -253,13 +284,30 @@ fn process_successful_execution<
             unimplemented!("Processing user events")
         }
     }
+    if gas_used > gas_budget {
+        return Err(FastPayError::InsufficientGas {
+            error: format!(
+                "Gas budget is {}, not enough to pay for cost {}",
+                gas_budget, gas_used
+            ),
+        });
+    }
     // any object left in `by_value_objects` is an input passed by value that was not transferred or frozen.
     // this means that either the object was (1) deleted from the FastX system altogether, or
     // (2) wrapped inside another object that is in the FastX object pool
     // in either case, we want to delete it
-    for id in by_value_objects.keys() {
+    let mut gas_refund: u64 = 0;
+    for (id, object) in by_value_objects.iter() {
         state_view.delete_object(id);
+        gas_refund += calculate_object_deletion_refund(object);
     }
+
+    // TODO: In the current approach, we basically can use refund gas to pay for current transaction.
+    // Is this allowed?
+    deduct_gas(&mut gas_object, (gas_used as i128) - (gas_refund as i128))?;
+    state_view.write_object(gas_object);
+    // TODO: Keep track the gas deducted so that we could give them to participants.
+
     Ok(())
 }
 
@@ -366,7 +414,7 @@ fn resolve_and_type_check(
     let mut args = Vec::new();
     let mut mutable_ref_objects = Vec::new();
     let mut by_value_objects = BTreeMap::new();
-    #[allow(unused_mut)]
+    #[cfg(debug_assertions)]
     let mut num_immutable_objects = 0;
     #[cfg(debug_assertions)]
     let num_objects = object_args.len();
